move verification if 1Tb nodes to sanity check
move verification if 1Tb nodes to sanity check
