+++ tests/virt/node/descheduler/test_descheduler.py
@@ -18,7 +18,6 @@ pytestmark = [
     pytest.mark.descheduler,
     pytest.mark.post_upgrade,
     pytest.mark.usefixtures(
-        "xfail_if_1tb_memory_or_more_node",
         "descheduler_long_lifecycle_profile",
     ),
 ]
