+++ tests/virt/node/descheduler/test_descheduler_psi_metrics.py
@@ -11,7 +11,6 @@ pytestmark = [
     pytest.mark.descheduler,
     pytest.mark.post_upgrade,
     pytest.mark.usefixtures(
-        "xfail_if_1tb_memory_or_more_node",
         "descheduler_kubevirt_relieve_and_migrate_profile",
     ),
 ]
